remote_agents/azurefoundry_QuickBooks/pyproject.toml
[project]
name = "aifoundry-quickbooks-agent"
version = "0.1.0"
description = "AI Foundry QuickBooks Online Agent with A2A protocol support"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk[http-server]>=0.2.6",
    "azure-ai-projects>=1.0.0b12",
    "azure-ai-agents>=1.1.0b2",
    "azure-identity>=1.23.0",
    "uvicorn>=0.34.2",
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
    "starlette>=0.35.0",
    "httpx>=0.25.0",
    "gradio>=4.0.0",
    "pandas>=2.0.0",
]
